SOLID Princip Opdatering ift. MainWindow og AddOrder
